requirements_fixed.txt
# Fixed requirements with compatible versions
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.13
langchain-google-genai==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
flask==3.0.0
flask-cors==4.0.0
fpdf2==2.7.6
pydantic==2.5.3
